feat: tidy constants and types
