+++ src/connect/constants/documents.ts
@@ -1,0 +1,15 @@
+export const LEGAL = 'legal';
+export const GLOBAL = 'global';
+export const PROFILE = 'profile';
+export const BUDGET = 'budget';
+export const TEMPLATE = 'template';
+export const MAKERDAO_RWA_PORTFOLIO = 'makerdao/rwa-portfolio';
+
+export const documentTypes = [
+    LEGAL,
+    GLOBAL,
+    PROFILE,
+    BUDGET,
+    TEMPLATE,
+    MAKERDAO_RWA_PORTFOLIO,
+] as const;
