+++ src/connect/constants/index.ts
@@ -1,1 +1,6 @@
-export * from './ui';
+export * from './documents';
+export * from './drives';
+export * from './icons';
+export * from './nodes';
+export * from './options';
+export * from './syncing';
