+++ src/connect/constants/nodes.ts
@@ -1,0 +1,6 @@
+export const DRIVE = 'DRIVE';
+export const FOLDER = 'FOLDER';
+export const FILE = 'FILE';
+export const CREATE = 'CREATE';
+export const WRITE = 'WRITE';
+export const READ = 'READ';
