+++ src/connect/types/documents.ts
@@ -1,0 +1,5 @@
+import { documentTypes } from '@/connect';
+
+export type DocumentTypes = typeof documentTypes;
+
+export type DocumentType = DocumentTypes[number];
