+++ src/connect/types/index.ts
@@ -1,1 +1,6 @@
-export * from './ui';
+export * from './documents';
+export * from './drives';
+export * from './nodes';
+export * from './options';
+export * from './shared-props';
+export * from './syncing';
