+++ src/connect/types/syncing.ts
@@ -1,0 +1,4 @@
+import { syncStatuses } from '@/connect';
+
+export type SyncStatuses = typeof syncStatuses;
+export type SyncStatus = SyncStatuses[number];
