+++ src/connect/utils/index.ts
@@ -2,5 +2,3 @@ export * from './address';
 export * from './get-folder-status';
 export * from './mocks';
 export * from './node-options';
-export * from './path';
-export * from './tree-item';
